+++ airflow/providers/jenking/hooks/__init__.py
@@ -16,14 +16,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.providers.jenking.hooks.jenkins`."""
-
-import warnings
-
-# pylint: disable=unused-import
-from airflow.providers.jenking.hooks.jenkins import JenkinsHook  # noqa
-
-warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.jenking.hooks.jenkins`.",
-    DeprecationWarning, stacklevel=2
-)
